prototype of check_label_utils extraction
