+++ .github/scripts/check_labels.py
@@ -7,11 +7,14 @@ from gitutils import (
     GitRepo,
 )
 from trymerge import (
+    gh_post_pr_comment,
+    GitHubPR,
+)
+from check_labels_utils import (
     LABEL_ERR_MSG,
     has_required_labels,
-    add_label_err_comment,
+    is_label_err_comment,
     delete_all_label_err_comments,
-    GitHubPR,
 )
 
 def parse_args() -> Any:
@@ -21,6 +24,10 @@ def parse_args() -> Any:
 
     return parser.parse_args()
 
+def add_label_err_comment(pr: GitHubPR) -> None:
+    # Only make a comment if one doesn't exist already
+    if not any(is_label_err_comment(comment) for comment in pr.get_comments()):
+        gh_post_pr_comment(pr.org, pr.project, pr.pr_num, LABEL_ERR_MSG)
 
 def main() -> None:
     args = parse_args()
@@ -29,11 +36,11 @@ def main() -> None:
     pr = GitHubPR(org, project, args.pr_num)
 
     try:
-        if not has_required_labels(pr):
+        if not has_required_labels(pr.get_labels(), pr.org, pr.project):
             print(LABEL_ERR_MSG)
             add_label_err_comment(pr)
         else:
-            delete_all_label_err_comments(pr)
+            delete_all_label_err_comments(pr.get_comments(), pr.org, pr.project)
     except Exception as e:
         pass
