+++ .github/scripts/trymerge.py
@@ -18,27 +18,30 @@ from typing import (
     List,
     NamedTuple,
     Optional,
-    Pattern,
     Tuple,
     Union,
     cast,
 )
-from urllib.error import HTTPError
-from urllib.request import Request, urlopen
 from warnings import warn
 from pathlib import Path
 
 from gitutils import (
     GitRepo,
+    GitHubComment,
     are_ghstack_branches_in_sync,
     get_git_remote_name,
     get_git_repo_dir,
     patterns_to_regex,
-    gh_get_labels,
+    gh_fetch_url,
 )
 from trymerge_explainer import (
     TryMergeExplainer,
     get_revert_message,
+)
+from check_labels_utils import (
+    has_required_labels,
+    delete_all_label_err_comments,
+    LABEL_ERR_MSG,
 )
 
 class JobCheckState(NamedTuple):
@@ -421,41 +424,8 @@ RE_PULL_REQUEST_RESOLVED = re.compile(
 )
 RE_PR_CC_LINE = re.compile(r'^cc:? @\w+.*\r?\n?$', re.MULTILINE)
 RE_DIFF_REV = re.compile(r'^Differential Revision:.+?(D[0-9]+)', re.MULTILINE)
-CIFLOW_LABEL = re.compile(r"^ciflow/.+")
-CIFLOW_TRUNK_LABEL = re.compile(r"^ciflow/trunk")
 MERGE_RULE_PATH = Path(".github") / "merge_rules.yaml"
 
-BOT_AUTHORS = ["github-actions", "pytorchmergebot", "pytorch-bot"]
-
-LABEL_ERR_MSG_TITLE = "This PR needs a label"
-LABEL_ERR_MSG = f"""# {LABEL_ERR_MSG_TITLE}\n
-    If your changes are user facing and intended to be a part of release notes, please use a label starting with `release notes:`.
-
-
-    If not, please add the `topic: not user facing` label.
-    For more information, see
-    https://github.com/pytorch/pytorch/wiki/PyTorch-AutoLabel-Bot#why-categorize-for-release-notes-and-how-does-it-work.
-"""
-
-
-def _fetch_url(url: str, *,
-               headers: Optional[Dict[str, str]] = None,
-               data: Optional[Dict[str, Any]] = None,
-               method: Optional[str] = None,
-               reader: Callable[[Any], Any] = lambda x: x.read()) -> Any:
-    if headers is None:
-        headers = {}
-    token = os.environ.get("GITHUB_TOKEN")
-    if token is not None and url.startswith('https://api.github.com/'):
-        headers['Authorization'] = f'token {token}'
-    data_ = json.dumps(data).encode() if data is not None else None
-    try:
-        with urlopen(Request(url, headers=headers, data=data_, method=method)) as conn:
-            return reader(conn)
-    except HTTPError as err:
-        if err.code == 403 and all(key in err.headers for key in ['X-RateLimit-Limit', 'X-RateLimit-Used']):
-            print(f"Rate limit exceeded: {err.headers['X-RateLimit-Used']}/{err.headers['X-RateLimit-Limit']}")
-        raise
 
 def fetch_json(url: str,
                params: Optional[Dict[str, Any]] = None,
@@ -463,7 +433,7 @@ def fetch_json(url: str,
     headers = {'Accept': 'application/vnd.github.v3+json'}
     if params is not None and len(params) > 0:
         url += '?' + '&'.join(f"{name}={urllib.parse.quote(str(val))}" for name, val in params.items())
-    return cast(List[Dict[str, Any]], _fetch_url(url, headers=headers, data=data, reader=json.load))
+    return cast(List[Dict[str, Any]], gh_fetch_url(url, headers=headers, data=data, reader=json.load))
 
 def fetch_json_dict(url: str,
                     params: Optional[Dict[str, Any]] = None,
@@ -471,7 +441,7 @@ def fetch_json_dict(url: str,
     headers = {'Accept': 'application/vnd.github.v3+json'}
     if params is not None and len(params) > 0:
         url += '?' + '&'.join(f"{name}={urllib.parse.quote(str(val))}" for name, val in params.items())
-    return cast(Dict[str, Any], _fetch_url(url, headers=headers, data=data, reader=json.load))
+    return cast(Dict[str, Any], gh_fetch_url(url, headers=headers, data=data, reader=json.load))
 
 def _gh_post_comment(url: str, comment: str, dry_run: bool = False) -> List[Dict[str, Any]]:
     if dry_run:
@@ -494,7 +464,7 @@ def gh_add_labels(org: str, project: str, pr_num: int, labels: Union[str, List[s
 
 
 def gh_graphql(query: str, **kwargs: Any) -> Dict[str, Any]:
-    rc = _fetch_url("https://api.github.com/graphql", data={"query": query, "variables": kwargs}, reader=json.load)
+    rc = gh_fetch_url("https://api.github.com/graphql", data={"query": query, "variables": kwargs}, reader=json.load)
     if "errors" in rc:
         raise RuntimeError(f"GraphQL query {query}, args {kwargs} failed: {rc['errors']}")
     return cast(Dict[str, Any], rc)
@@ -670,15 +640,6 @@ def get_ghstack_prs(repo: GitRepo, pr: "GitHubPR") -> List[Tuple["GitHubPR", str
                 f"Please sync them and try again (ex. make the changes on {orig_ref} and run ghstack)."
             )
     return entire_stack
-
-@dataclass
-class GitHubComment:
-    body_text: str
-    created_at: str
-    author_login: str
-    author_association: str
-    editor_login: Optional[str]
-    database_id: int
 
 
 class GitHubPR:
@@ -1109,7 +1070,7 @@ def gen_new_issue_link(
 def read_merge_rules(repo: Optional[GitRepo], org: str, project: str) -> List[MergeRule]:
     repo_relative_rules_path = MERGE_RULE_PATH
     if repo is None:
-        json_data = _fetch_url(
+        json_data = gh_fetch_url(
             f"https://api.github.com/repos/{org}/{project}/contents/{repo_relative_rules_path}",
             headers={'Accept': 'application/vnd.github.v3+json'},
             reader=json.load,
@@ -1408,37 +1369,6 @@ def validate_land_time_checks(org: str, project: str, commit: str) -> None:
     if len(pending_checks) > 0:
         raise MandatoryChecksMissingError(f"Refusing to merge as land check(s) {checks_to_str(pending_checks)} are not yet run")
 
-def has_label(labels: List[str], pattern: Pattern[str] = CIFLOW_LABEL) -> bool:
-    return len(list(filter(pattern.match, labels))) > 0
-
-def get_release_notes_labels(org: str, repo: str) -> List[str]:
-    return [label for label in gh_get_labels(org, repo) if label.lstrip().startswith("release notes:")]
-
-def has_required_labels(pr: GitHubPR) -> bool:
-    pr_labels = pr.get_labels()
-    # Check if PR is not user facing
-    is_not_user_facing_pr = any(label.strip() == "topic: not user facing" for label in pr_labels)
-    return (
-        is_not_user_facing_pr or
-        any(label.strip() in get_release_notes_labels(pr.org, pr.project) for label in pr_labels)
-    )
-
-def gh_post_delete_comment(org: str, project: str, comment_id: int) -> None:
-    url = f"https://api.github.com/repos/{org}/{project}/issues/comments/{comment_id}"
-    _fetch_url(url, method="DELETE")
-
-def is_label_err_comment(comment: GitHubComment) -> bool:
-    return comment.body_text.lstrip(" #").startswith(LABEL_ERR_MSG_TITLE) and comment.author_login in BOT_AUTHORS
-
-def delete_all_label_err_comments(pr: GitHubPR) -> None:
-    for comment in pr.get_comments():
-        if is_label_err_comment(comment):
-            gh_post_delete_comment(pr.org, pr.project, comment.database_id)
-
-def add_label_err_comment(pr: GitHubPR) -> None:
-    # Only make a comment if one doesn't exist already
-    if not any(is_label_err_comment(comment) for comment in pr.get_comments()):
-        gh_post_pr_comment(pr.org, pr.project, pr.pr_num, LABEL_ERR_MSG)
 
 def categorize_checks(check_runs: Dict[str, JobCheckState],
                       required_checks: Iterable[str]) -> Tuple[List[Tuple[str, Optional[str]]], List[Tuple[str, Optional[str]]]]:
@@ -1497,10 +1427,10 @@ def merge(pr_num: int, repo: GitRepo,
     # here to stop the merge process right away
     find_matching_merge_rule(pr, repo, skip_mandatory_checks=True)
 
-    if not has_required_labels(pr):
+    if not has_required_labels(pr.get_labels(), pr.org, pr.project):
         raise RuntimeError(LABEL_ERR_MSG.lstrip(" #"))
     else:
-        delete_all_label_err_comments(pr)
+        delete_all_label_err_comments(pr.get_comments(), pr.org, pr.project)
 
     if land_checks and not dry_run:
         land_check_commit = pr.create_land_time_check_branch(
